% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAI.R
\name{gap_fraction}
\alias{gap_fraction}
\title{Calculates the gap fraction (ratio of open canopy) from a binary image}
\usage{
gap_fraction(binary_img)
}
\arguments{
\item{binary_img}{A binarized raster object, where sky pixels are 0 and
vegetation pixels are 1}
}
\value{
The gap fraction ( number of sky pixels / total number of pixels )
}
\description{
Calculates the gap fraction (ratio of open canopy) from a binary image
}
\examples{
library(raster)
gap_fraction(
 unimodal_threshold(
   raster(system.file("extdata", "IMG_0005.JPG", package = "LAI"), band = 3)
 )
)
}
\seealso{
\code{\link{unimodal_threshold}} \code{\link{LAI_from_gf_at_57}}
}

