\name{LAI-package}
\alias{LAI-package}
\alias{LAI}
\docType{package}
\title{
\packageTitle{LAI}
}
\description{
\packageDescription{LAI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LAI}
\packageIndices{LAI}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LAI}

Maintainer: \packageMaintainer{LAI}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
