% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LAI.R
\name{LAI_from_gf_at_57}
\alias{LAI_from_gf_at_57}
\title{Calculate leaf area index (LAI) from an image in a band around 57.5 degrees.}
\usage{
LAI_from_gf_at_57(image_path)
}
\arguments{
\item{image_path}{Path to the image to analyze.}
}
\value{
The calculated LAI value.
}
\description{
Calculate leaf area index (LAI) from an image in a band around 57.5 degrees.
}
\examples{
LAI_from_gf_at_57("IMG_0005.JPG")
}

