% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{unimodal_threhsold}
\alias{unimodal_threhsold}
\title{Binarize a vegetation image between sky and vegetation}
\usage{
unimodal_threhsold(img)
}
\description{
Binarize a vegetation image between sky and vegetation
}

