% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unimodal_threshold}
\alias{unimodal_threshold}
\title{Binarize an image between sky and vegetation pixels}
\usage{
unimodal_threshold(img)
}
\arguments{
\item{img}{A raster object with a single (blue) band.}
}
\value{
A raster object, where sky pixels are 0 and vegetation pixels are 1.
}
\description{
This functions applies the histogram-based unimodal thresholding method
}
\examples{
library(raster)
file_path = system.file("extdata", "IMG_7595.JPG", package = "LAI")

# View the original image
plotRGB(
  brick(file_path),
  asp = 1
)

# View the binarized image
plot(
  unimodal_threshold(
    raster(file_path, band = 3)
  )
  ,useRaster = TRUE
  ,asp = 1
)
}
\references{
Macfarlane, C. (2011). Classification method of mixed pixels does not affect
canopy metrics from digital images of forest overstorey. Agricultural and
Forest Meteorology, 151(7), 833-840.
}
\seealso{
\code{\link{gap_fraction}} \code{\link{LAI_from_gf_at_57}}
}

