% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LAI.R
\name{LAI_from_gf_at_57}
\alias{LAI_from_gf_at_57}
\title{Calculate leaf area index (LAI) from an image in a band around 57.5 degrees.}
\usage{
LAI_from_gf_at_57(image_path, camera_horiz_FOV = 73.7, focal_angle = 45)
}
\arguments{
\item{image_path}{Path to the image to analyze.}

\item{camera_horiz_FOV}{Camera horizontal field of view (in degrees)}

\item{focal_angle}{Angle at which the camera was pointing (degrees). 0 is
horizontal, 90 is vertical}
}
\value{
The calculated LAI value.
}
\description{
This functions first seperates sky from vegetation pixels using
\code{\link{unimodal_threshold}}. Based on the
focal angle and camera field of view provided, it then extracts a narrow
band around a 57.5 degrees zenith angle (Baret et al. 2010). The
\code{\link{gap_fraction}} of this band is then used to
calculated an indirect LAI value (e.g. Confalonieri et al. 2013).
}
\examples{
LAI_from_gf_at_57(
  system.file("extdata", "IMG_7595.JPG", package = "LAI")
)
}
\references{
Baret, F., de Solan, B., Lopez-Lozano, R., Ma, K., & Weiss, M. (2010).
GAI estimates of row crops from downward looking digital photos taken
perpendicular to rows at 57.5 zenith angle: Theoretical considerations
based on 3D architecture models and application to wheat crops.
Agricultural and Forest Meteorology, 150(11), 1393-1401.

  Confalonieri, R., Foi, M., Casa, R., Aquaro, S., Tona, E., Peterle,
M., ... Acutis, M. (2013). Development of an app for estimating leaf area
index using a smartphone. Trueness and precision determination and
comparison with other indirect methods. Computers and Electronics in
Agriculture, 96, 67-74.
}

